refactor #31: user 정보 조회를 위한 비지니스 로직 hook 분리
