+++ src/models/index.ts
@@ -1,0 +1,7 @@
+const QUERY_KEY = ['user'];
+
+const cache = {
+  getUser: () => [QUERY_KEY],
+};
+
+export default cache;
