--- src/queries/index.ts
@@ -1,1 +1,0 @@
-export { default as useUser } from './useUserQuery';
