--- src/queries/useUserQuery/types.ts
@@ -1,7 +1,0 @@
-export interface UserInfo {
-  nickname: string;
-  socialType: string;
-  thumbnailUrl: string;
-  uid: number;
-  username: string;
-}
